+++ src/rules/declaration-property-value-whitelist/index.js
@@ -7,7 +7,7 @@ import {
   matchesStringOrRegExp,
 } from "../../utils"
 
-export const ruleName = "property-value-whitelist"
+export const ruleName = "declaration-property-value-whitelist"
 
 export const messages = ruleMessages(ruleName, {
   rejected: (property, value) => `Unexpected value "${value}" for property "${property}"`,
@@ -15,15 +15,6 @@ export const messages = ruleMessages(ruleName, {
 
 export default function (whitelist) {
   return (root, result) => {
-
-    result.warn((
-      "'property-value-whitelist' has been deprecated, "
-        + "and will be removed in '7.0'. Use 'declaration-property-value-whitelist' instead."
-    ), {
-      stylelintType: "deprecation",
-      stylelintReference: "http://stylelint.io/user-guide/rules/declaration-property-value-whitelist/",
-    })
-
     const validOptions = validateOptions(result, ruleName, {
       actual: whitelist,
       possible: [isObject],