+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Head
 
 - Deprecated: `stylelint.util.styleSearch()`. Use the external module [style-search](https://github.com/davidtheclark/style-search) instead.
+- Deprecated: `property-*-blacklist` and `property-*-whitelist` rules. Use `declaration-property-*-blacklist` and `declaration-property-*-whitelist` instead.
 - Added: option `ignorePath` (for JS) and `--ignore-path` (for CLI).
 - Added: `at-rule-name-newline-after` rule.
 - Added: `at-rule-blacklist` rule.